requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
requests==2.31.0
pyserial==3.5
Werkzeug==3.0.1
SQLAlchemy==2.0.23

# Sensor libraries (optional - for hardware sensors)
adafruit-circuitpython-bme680
adafruit-circuitpython-veml6075

# SMS/Call integration (optional)
twilio

# Additional utilities
python-dateutil
